% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clic.R
\name{clic}
\alias{clic}
\title{Composite likelihood Information Criterion}
\usage{
clic(logCL, mat.hessian, mat.J)
}
\arguments{
\item{logCL}{value of composite log-likelihood.}

\item{mat.hessian}{hessian matrix.}

\item{mat.J}{Sensitivity matrix}
}
\value{
\code{CLIC}: Composite likelihood information criterion proposed by Varin and Vidoni (2005)

\code{clic}: Composite likelihood information criterion proposed by Varin and Vidoni (2005)
}
\description{
\code{clic} Calculating the Composite likelihood information criterion proposed by Varin and Vidoni (2005)
}
\details{
Varin and Vidoni (2005) proposed the information criterion in the form:
\eqn{-2*logCL(theta) + 2*trace(H^{-1}(\theta)J(\theta))}
}
\references{
Varin, C. and Vidoni, P. (2005) A note on composite likelihood inference and model selection. \emph{Biometrika} 92: 519--528.
}
