% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged.list.R
\name{merge.list}
\alias{merge.list}
\title{Composite Likelihood Calculation for Spatial Ordinal Data without Replications (for implmentation)}
\usage{
\method{merge}{list}(x, y = NULL, mergeUnnamed = TRUE)
}
\arguments{
\item{x}{an object to be merged into list of object.}

\item{y}{an object to be merged into list.}

\item{mergeUnnamed}{select an element if it has a.) an empty name in y and
mergeUnnamed is true or  b.) a name _not_ contained in x}
}
\value{
\code{merge.list} returns a list: a merged list
}
\description{
Composite Likelihood Calculation for Spatial Ordinal Data without Replications (for implmentation)
}
